requirements.txt
flask>=3.0.0
flask-cors>=6.0.0
flask-socketio==5.3.6
gunicorn>=23.0.0
openai>=2.0.0
pillow>=10.0.0
playwright>=1.40.0
python-socketio==5.11.0
websocket-client>=1.0.0
gevent>=25.0.0
gevent-websocket>=0.10.0
eventlet>=0.40.0
